Genex: Add $<HAVE_COMPILER_FEATURE:FEATURE> genex.
Implement a consistency check similar to the existing check for
target properties seen when evaluating the link implementation.
A compile feature may not both determine and be determined by
the link implementation.

Record a set of features which are available by default for
a compiler if not features are requested explicitly for a target.