Windows: Use response files for link libraries only where supported
The NMake Makefiles generator already puts link libraries in an inline
response file and at least the VS <= 7.1 tools do not support nested
response files.  GNU 3.x compilers do not support response files so we
pass them to the linker using -Wl,@ but link libraries are generated
with options meant for the compiler front-end so passing them through
a response file to the linker does not always work.