Makefile: Generate single-quoted linker lines for Watcom on Windows
Drop the CMAKE_NO_QUOTED_OBJECTS internal variable from the Makefile
generators.  The underlying problem is with the Watcom linker, not with
WMake.  The Watcom linker wants object files to be single-quoted since
it does its own parsing of the Windows process command-line string.  Add
<LINK-RULE>_USE_SINGLE_QUOTE platform information variables to tell the
generators to use single quotes for object files on link lines.