Export: Disallow exported interface includes in src/build tree (#14592).
Allow directories in the build tree only if they are also
subdirectories of the install tree.

Re-use the test files in the RunCMake.include_directories test
to run in multiple scenarios.  Bump the required CMake version
in the test to 3.0 to ensure that the new policy warnings are
emitted correctly.