QT4_CREATE_MOC_COMMAND is now a function.
When we parse a large string ( moc_flags ) that has generator expressions
while inside a macro it has poor performance. This is due to the fact
that the because we are in a macro we have to evaluate the contents of the
string. By moving to a function we eliminate the need to evaluate the entire
contents of the string, and increase the performance.

In non scientific testing I have seen a 10x performance increase by moving
QT4_CREATE_MOC_COMMAND to being a function.