HP aCC: record compiler flag for Ansi C++98 support for version 3.80 onwards
Old versions of aCC need a special compiler flag to get full C++98 template
support as e.g. CMake itself or the Complex and ComplexOneConfig tests need.

For merge simplicity this patch is based on master since it doesn't collide
with anything from the cxx11-features branch, which this feature depends on.
This branch should only go into master after the CXX_STANDARD variable is
merged to master, and this paragraph should be removed from the commit message
then.